requirements.txt
streamlit==1.31.1
google-genai==0.2.2
python-dotenv==1.0.0
